+++ development/app/Http/Controllers/Backend/CustomerController.php
@@ -73,7 +73,11 @@ class CustomerController extends Controller
     }
     public function destroy($id){
         $customer = User::find($id);
+        if (!empty($customer)){
         $customer->delete();
+        }else{
+
+        }
         return redirect()->route('customers.index');
     }
 }
